+++ src/components/HomeGallery/HomeGalleryFilter/HomeGalleryFilterTypes.module.css
@@ -1,0 +1,3 @@
+.formGroupElement {
+  margin-bottom: 0.75rem;
+}
