+++ src/components/HomeGallery/HomeGalleryFilter/index.ts
@@ -1,1 +1,4 @@
 export * from './HomeGalleryFilter';
+export * from './HomeGalleryFilterSourceRepo';
+export * from './HomeGalleryFilterTags';
+export * from './HomeGalleryFilterTypes';
